Add return User , delete Project, token all routes

+++ src/controllers/ProjectController.ts
@@ -6,6 +6,11 @@ export class ProjectController {
     static createProject = async (req: Request, res: Response) => {
 
         const project = new Project(req.body)
+
+        // Asigna un manager
+        project.manager = req.user.id
+
+
         try {
             await project.save()
             res.send('Proyecto creado correctamente')
@@ -18,7 +23,11 @@ export class ProjectController {
 
     static getAllProjects = async (req: Request, res: Response) => {
         try {
-            const projects = await Project.find({})
+            const projects = await Project.find({
+                $or: [
+                    {manager: {$in: req.user.id}}
+                ]
+            })
             res.json(projects)
         } catch (error) {
             console.log(error);
@@ -32,6 +41,10 @@ export class ProjectController {
         try {
             const project = await Project.findById(id).populate('tasks')
 
+            if(project.manager.toString() !== req.user.id.toString()){
+                const error =new Error('Acción no valida')
+                return res.status(404).json({error: error.message})
+            }
             res.json(project)
         } catch (error) {
             console.log(error);
@@ -44,6 +57,12 @@ export class ProjectController {
 
         try {
             const project = await Project.findById(id)
+
+            if(project.manager.toString() !== req.user.id.toString()){
+                const error =new Error('Solo el manager puede actualizar un proyecto')
+                return res.status(404).json({error: error.message})
+            }
+
             project.clientName = req.body.clientName
             project.projectName = req.body.projectName
             project.description = req.body.description
@@ -60,6 +79,10 @@ export class ProjectController {
 
         try {
             const project = await Project.findById(id)
+            if(project.manager.toString() !== req.user.id.toString()){
+                const error =new Error('Solo el manager puede eliminar un proyecto')
+                return res.status(404).json({error: error.message})
+            }
             await project.deleteOne()
             res.send('Proyecto eliminado')
         } catch (error) {
